Prophecy
This is a project which checks whether a person is willing to buy a bike in future based on his/her some demographic informations. Several preprocessing were done on the dataset to deal with the missing values, outliers and for normalisation. Support Vector Machine and decision tree are the predictive models applied on the dataset and which gives an accuracy close to 90%.
UI : Flask